requirements.txt
streamlit
psycopg2-binary
sqlalchemy
langchain
langchain-community
ollama
python-dotenv
chromadb
passlib[bcrypt]
bcrypt==4.0.1
langchain-ollama
fastapi
uvicorn[standard]
python-jose[cryptography]
requests
python-multipart
altair
pandas
